+++ AudioKit/OSX/AudioKit/Playgrounds/Helpers/AKDynamicsProcessorWindow.swift
@@ -191,52 +191,52 @@ public class AKDynamicsProcessorWindow: NSWindow {
     }
     
     internal func updateThreshold() {
-        dynamicsProcessor.threshold = thresholdSlider.floatValue
+        dynamicsProcessor.threshold = thresholdSlider.doubleValue
         thresholdTextField.stringValue =
         "Threshold \(String(format: "%0.4f", dynamicsProcessor.threshold)) dB"
     }
     internal func updateHeadroom() {
-        dynamicsProcessor.headRoom = headRoomSlider.floatValue
+        dynamicsProcessor.headRoom = headRoomSlider.doubleValue
         headRoomTextField.stringValue =
         "Head Room \(String(format: "%0.4f", dynamicsProcessor.headRoom)) dB"
     }
     internal func updateExpansionratio() {
-        dynamicsProcessor.expansionRatio = expansionRatioSlider.floatValue
+        dynamicsProcessor.expansionRatio = expansionRatioSlider.doubleValue
         expansionRatioTextField.stringValue =
         "Expansion Ratio \(String(format: "%0.4f", dynamicsProcessor.expansionRatio)) rate"
     }
     internal func updateExpansionthreshold() {
-        dynamicsProcessor.expansionThreshold = expansionThresholdSlider.floatValue
+        dynamicsProcessor.expansionThreshold = expansionThresholdSlider.doubleValue
         expansionThresholdTextField.stringValue =
         "Expansion Threshold \(String(format: "%0.4f", dynamicsProcessor.expansionThreshold)) rate"
     }
     internal func updateAttacktime() {
-        dynamicsProcessor.attackTime = attackTimeSlider.floatValue
+        dynamicsProcessor.attackTime = attackTimeSlider.doubleValue
         attackTimeTextField.stringValue =
         "Attack Time \(String(format: "%0.4f", dynamicsProcessor.attackTime)) secs"
     }
     internal func updateReleasetime() {
-        dynamicsProcessor.releaseTime = releaseTimeSlider.floatValue
+        dynamicsProcessor.releaseTime = releaseTimeSlider.doubleValue
         releaseTimeTextField.stringValue =
         "Release Time \(String(format: "%0.4f", dynamicsProcessor.releaseTime)) secs"
     }
     internal func updateMastergain() {
-        dynamicsProcessor.masterGain = masterGainSlider.floatValue
+        dynamicsProcessor.masterGain = masterGainSlider.doubleValue
         masterGainTextField.stringValue =
         "Master Gain \(String(format: "%0.4f", dynamicsProcessor.masterGain)) dB"
     }
     internal func updateCompressionamount() {
-        dynamicsProcessor.compressionAmount = compressionAmountSlider.floatValue
+        dynamicsProcessor.compressionAmount = compressionAmountSlider.doubleValue
         compressionAmountTextField.stringValue =
         "Compression Amount \(String(format: "%0.4f", dynamicsProcessor.compressionAmount)) dB"
     }
     internal func updateInputamplitude() {
-        dynamicsProcessor.inputAmplitude = inputAmplitudeSlider.floatValue
+        dynamicsProcessor.inputAmplitude = inputAmplitudeSlider.doubleValue
         inputAmplitudeTextField.stringValue =
         "Input Amplitude \(String(format: "%0.4f", dynamicsProcessor.inputAmplitude)) dB"
     }
     internal func updateOutputamplitude() {
-        dynamicsProcessor.outputAmplitude = outputAmplitudeSlider.floatValue
+        dynamicsProcessor.outputAmplitude = outputAmplitudeSlider.doubleValue
         outputAmplitudeTextField.stringValue =
         "Output Amplitude \(String(format: "%0.4f", dynamicsProcessor.outputAmplitude)) dB"
     }
